+++ lib/class/Containers/ControllerContainer/BodyController/index.d.ts
@@ -3,11 +3,17 @@ import { IConstructorBaseProps } from '../../../../type';
 interface IConstructorProps extends IConstructorBaseProps {
 }
 declare class BodyController extends BaseComponent {
-    private buttonPrimary;
+    private buttonPlayPrimary;
+    private buttonReplayPrimary;
     private settingsController;
     constructor(props: IConstructorProps);
     render(): void;
     destroy(): void;
+    registerListener(): void;
+    unregisterListener(): void;
+    handleEventPlay(): void;
+    handleEventPause(): void;
+    handleEventEnded(): void;
 }
 export default BodyController;
 //# sourceMappingURL=index.d.ts.map