+++ lib/class/Containers/ControllerContainer/FooterController/TaskbarController/VolumeContainer/index.d.ts
@@ -11,6 +11,9 @@ declare class VolumeContainer extends BaseComponent<IConstructorProps> {
     render(): void;
     registerListener(): void;
     unregisterListener(): void;
+    handelEventClick(e: MouseEvent): void;
+    handelEventMouseover(e: MouseEvent): void;
+    handelEventMouseout(e: MouseEvent): void;
 }
 export default VolumeContainer;
 //# sourceMappingURL=index.d.ts.map