+++ lib/class/Containers/ControllerContainer/FooterController/TaskbarController/index.d.ts
@@ -6,6 +6,7 @@ declare class TaskbarController extends BaseComponent {
     private buttonFullScreen;
     private buttonPlaySecondary;
     private buttonPauseSecondary;
+    private buttonReplaySecondary;
     private volumeContainer;
     private buttonExitFullScreen;
     private settingIconButton;
@@ -13,6 +14,13 @@ declare class TaskbarController extends BaseComponent {
     private liveStream;
     constructor(props: IConstructorProps);
     render(): void;
+    registerListener(): void;
+    unregisterListener(): void;
+    handleEventPlay(): void;
+    handleEventPause(): void;
+    handleEventLoaded(): void;
+    handleEventFullScreen(): void;
+    handleEventEnded(): void;
 }
 export default TaskbarController;
 //# sourceMappingURL=index.d.ts.map