+++ lib/constants.d.ts
@@ -19,7 +19,6 @@ export declare const ids: {
     smButtonPlaySecondary: string;
     smButtonPauseSecondary: string;
     smButtonForward: string;
-    smButtonReply: string;
     smButtonExitFullScreen: string;
     smButtonVolume: string;
     smButtonMute: string;
@@ -37,6 +36,8 @@ export declare const ids: {
     smProgressThumb: string;
     smProgressBarContainer: string;
     smProgressBuffer: string;
+    smButtonReplaySecondary: string;
+    smButtonReplayPrimary: string;
 };
 export declare const versionDef = "4.10.0";
 export declare enum ETypePlayer {